Date 관련 기능 정리, 파일명 변경
